fastapi_service/requirements.txt
fastapi
uvicorn
python-jose[cryptography]
requests
pydantic
psycopg2-binary
